Street_App/frontend/app/api/requirements.txt
fastapi==0.114.1
python-multipart==0.0.6
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.1
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.1.23
langchain-text-splitters==0.3.0
openai==1.3.0
PyPDF2==3.0.1 